move more exceptions, fix warn-errors flag
