+++ core/dbt/adapters/exceptions/connection.py
@@ -1,0 +1,17 @@
+from typing import List
+
+from dbt.common.exceptions import DbtRuntimeError
+from dbt.exceptions import DbtDatabaseError
+
+
+class InvalidConnectionError(DbtRuntimeError):
+    def __init__(self, thread_id, known: List) -> None:
+        self.thread_id = thread_id
+        self.known = known
+        super().__init__(
+            msg=f"connection never acquired for thread {self.thread_id}, have {self.known}"
+        )
+
+
+class FailedToConnectError(DbtDatabaseError):
+    pass
