+++ core/dbt/common/events/format.py
@@ -1,9 +1,9 @@
 from dbt.common import ui
 
-# ToDo Figure out how to import NodeType
-# from dbt.node_types import NodeType
 from typing import Optional, Union
 from datetime import datetime
+
+from dbt.common.events.interfaces import LoggableDbtObject
 
 
 def format_fancy_output_line(
@@ -37,19 +37,17 @@ def format_fancy_output_line(
     return output
 
 
-# def _pluralize(string: Union[str, NodeType]) -> str:
-#     try:
-#         convert = NodeType(string)
-#     except ValueError:
-#         return f"{string}s"
-#     else:
-#         return convert.pluralize()
+def _pluralize(string: Union[str, LoggableDbtObject]) -> str:
+    if isinstance(string, LoggableDbtObject):
+        return string.pluralize()
+    else:
+        return f"{string}s"
 
 
-def pluralize(count, string: Union[str]) -> str:
+def pluralize(count, string: Union[str, LoggableDbtObject]) -> str:
     pluralized: str = str(string)
-    # if count != 1:
-    #     pluralized = _pluralize(string)
+    if count != 1:
+        pluralized = _pluralize(string)
     return f"{count} {pluralized}"
