+++ core/dbt/common/events/interfaces.py
@@ -1,0 +1,7 @@
+from typing import Protocol, runtime_checkable
+
+
+@runtime_checkable
+class LoggableDbtObject(Protocol):
+    def pluralize(self) -> str:
+        ...
