+++ core/dbt/task/runnable.py
@@ -38,10 +38,10 @@ from dbt.common.events.types import (
 )
 from dbt.exceptions import (
     DbtInternalError,
-    NotImplementedError,
     DbtRuntimeError,
     FailFastError,
 )
+from ..common.exceptions import NotImplementedError
 from dbt.flags import get_flags
 from dbt.graph import GraphQueue, NodeSelector, SelectionSpec, parse_difference, UniqueId
 from dbt.logger import (
