+++ tests/adapter/dbt/tests/adapter/utils/test_validate_sql.py
@@ -3,7 +3,8 @@ from typing import Type
 import pytest
 
 from dbt.adapters.base.impl import BaseAdapter
-from dbt.exceptions import DbtRuntimeError, InvalidConnectionError
+from dbt.exceptions import DbtRuntimeError
+from dbt.adapters.exceptions.connection import InvalidConnectionError
 
 
 class BaseValidateSqlMethod:
