+++ tests/unit/test_adapter_connection_manager.py
@@ -2,6 +2,7 @@ import unittest
 from unittest import mock
 import sys
 
+import dbt.adapters.exceptions.connection
 import dbt.exceptions
 
 import psycopg2
@@ -75,7 +76,8 @@ class BaseConnectionManagerTest(unittest.TestCase):
             raise ValueError("Something went horribly wrong")
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError, "Something went horribly wrong"
+            dbt.adapters.exceptions.connection.FailedToConnectError,
+            "Something went horribly wrong",
         ):
 
             BaseConnectionManager.retry_connection(
@@ -110,7 +112,8 @@ class BaseConnectionManagerTest(unittest.TestCase):
             raise ValueError("Something went horribly wrong")
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError, "Something went horribly wrong"
+            dbt.adapters.exceptions.connection.FailedToConnectError,
+            "Something went horribly wrong",
         ):
 
             BaseConnectionManager.retry_connection(
@@ -185,7 +188,8 @@ class BaseConnectionManagerTest(unittest.TestCase):
             raise ValueError("Something went horribly wrong")
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError, "Something went horribly wrong"
+            dbt.adapters.exceptions.connection.FailedToConnectError,
+            "Something went horribly wrong",
         ):
             BaseConnectionManager.retry_connection(
                 conn,
@@ -220,7 +224,8 @@ class BaseConnectionManagerTest(unittest.TestCase):
             raise TypeError("An unhandled thing went horribly wrong")
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError, "An unhandled thing went horribly wrong"
+            dbt.adapters.exceptions.connection.FailedToConnectError,
+            "An unhandled thing went horribly wrong",
         ):
             BaseConnectionManager.retry_connection(
                 conn,
@@ -338,7 +343,8 @@ class BaseConnectionManagerTest(unittest.TestCase):
             return True
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError, "retry_limit cannot be negative"
+            dbt.adapters.exceptions.connection.FailedToConnectError,
+            "retry_limit cannot be negative",
         ):
             BaseConnectionManager.retry_connection(
                 conn,
@@ -365,7 +371,7 @@ class BaseConnectionManagerTest(unittest.TestCase):
 
         for retry_timeout in [-10, -2.5, lambda _: -100, lambda _: -10.1]:
             with self.assertRaisesRegex(
-                dbt.exceptions.FailedToConnectError,
+                dbt.adapters.exceptions.connection.FailedToConnectError,
                 "retry_timeout cannot be negative or return a negative time",
             ):
                 BaseConnectionManager.retry_connection(
@@ -392,7 +398,7 @@ class BaseConnectionManagerTest(unittest.TestCase):
             return True
 
         with self.assertRaisesRegex(
-            dbt.exceptions.FailedToConnectError,
+            dbt.adapters.exceptions.connection.FailedToConnectError,
             "retry_limit cannot be negative",
         ):
             BaseConnectionManager.retry_connection(
