+++ tests/unit/test_semver.py
@@ -2,7 +2,7 @@ import unittest
 import itertools
 
 from typing import List
-from dbt.exceptions import VersionsNotCompatibleError
+from dbt.common.exceptions import VersionsNotCompatibleError
 from dbt.semver import (
     VersionSpecifier,
     UnboundedVersionSpecifier,
